+++ vllm/entrypoints/pooling/__init__.py
@@ -10,7 +10,7 @@ def register_pooling_api_routers(app: FastAPI):
     from vllm.entrypoints.pooling.pooling.api_router import router as pooling_router
     from vllm.entrypoints.pooling.score.api_router import router as score_router
 
-    app.include_router(classify_router)
-    app.include_router(embed_router)
-    app.include_router(score_router)
-    app.include_router(pooling_router)
+    # app.include_router(classify_router)
+    # app.include_router(embed_router)
+    # app.include_router(score_router)
+    # app.include_router(pooling_router)
